refactor(utils): move concise functions to utils
